test(parser): remove codegen dev dependency
