+++ crates/swc_ecma_parser/Cargo.toml
@@ -60,7 +60,6 @@ swc_ecma_ast = { version = "29.0.0", path = "../swc_ecma_ast", features = [
   "serde-impl",
   "encoding-impl",
 ] }
-swc_ecma_codegen = { version = "32.0.0", path = "../swc_ecma_codegen" }
 swc_ecma_visit = { version = "29.0.0", path = "../swc_ecma_visit" }
 swc_malloc = { version = "1.2.5", path = "../swc_malloc" }
 testing = { version = "27.0.0", path = "../testing" }
